v2.0.0 - part of php8.2 and phpstan level 9

+++ src/Helpers.php
@@ -1,18 +1,20 @@
 <?php
 namespace CryCMS;
 
+use JetBrains\PhpStorm\NoReturn;
 use JsonException;
 
 class Helpers
 {
-    public static function aPre($data, $end = false): void
+    public static function aPre(mixed $data, bool $end = false): void
     {
         if (self::isJson($data)) {
             try {
+                /** @var string $data */
                 $data = json_decode($data, false, 512, JSON_THROW_ON_ERROR);
                 $data = json_encode($data, JSON_THROW_ON_ERROR | JSON_PRETTY_PRINT | JSON_UNESCAPED_UNICODE);
             } catch (JsonException $exception) {
-                $data = 'error json';
+                $data = 'error json: ' . $exception->getMessage();
             }
         }
 
@@ -30,17 +32,26 @@ class Helpers
 
         $trace = debug_backtrace();
 
-        echo "<pre>";
-        echo "<small>" . str_replace(DR, '', $trace[0]['file']) . ":" . $trace[0]['line'] . "</small>" . PHP_EOL;
-        print_r($data);
-        echo "</pre>" . PHP_EOL;
+        if (
+            array_key_exists(0, $trace) &&
+            array_key_exists('file', $trace[0]) &&
+            array_key_exists('line', $trace[0])
+        ) {
+            echo "<pre>";
+            echo "<small>" . str_replace(DR, '', $trace[0]['file']) . ":" . $trace[0]['line'] . "</small>" . PHP_EOL;
+            print_r($data);
+            echo "</pre>" . PHP_EOL;
+        }
 
         if ($end) {
             exit;
         }
     }
 
-    public static function isJson($string): bool
+    /**
+     * @phpstan-assert-if-true string $string
+     */
+    public static function isJson(mixed $string): bool
     {
         if (is_array($string)) {
             return false;
@@ -54,28 +65,29 @@ class Helpers
             return false;
         }
 
-        $ss = preg_replace('/"(\\.|[^"\\\\])*"/', '', $string);
-        if (preg_match('/[^,:{}\\[\\]0-9.\\-+Eaeflnr-u \\n\\r\\t]/', $ss) === false) {
-            return true;
-        }
+        if (is_string($string)) {
+            $ss = preg_replace('/"(\\.|[^"\\\\])*"/', '', $string);
+            if (!empty($ss) && preg_match('/[^,:{}\\[\\]0-9.\\-+Eaeflnr-u \\n\\r\\t]/', $ss) === false) {
+                return true;
+            }
 
-        try {
-            $json = json_decode($string, false, 512, JSON_THROW_ON_ERROR);
-            return $json && $string !== $json;
-        }
-        catch (JsonException $exception) {
+            try {
+                $json = json_decode($string, false, 512, JSON_THROW_ON_ERROR);
+                return $json && $string !== $json;
+            } catch (JsonException) {
+            }
         }
 
         return false;
     }
 
-    public static function clean($string, $max_length = false): string
+    public static function clean(string $string, ?int $maxLength = null): string
     {
         $string = strip_tags($string);
         $string = htmlentities($string, ENT_NOQUOTES, "UTF-8");
 
-        if (!empty($max_length)) {
-            $string = mb_substr($string, 0, $max_length, 'UTF-8');
+        if ($maxLength !== null) {
+            $string = mb_substr($string, 0, $maxLength, 'UTF-8');
         }
 
         return trim($string);
@@ -90,16 +102,22 @@ class Helpers
      */
     public static function removeDuplicateCharacter(string $string, string $char = '\s', string $replacement = ' '): string
     {
-        return preg_replace('/' . $char . '{2,}/', $replacement, $string);
+        return preg_replace('/' . $char . '{2,}/', $replacement, $string) ?? $string;
     }
 
-    public static function redirect($location): void
+    #[NoReturn]
+    public static function redirect(string $location): void
     {
         Header("Location: " . $location);
         exit;
     }
 
-    public static function buildFullPath($url, $get): string
+    /**
+     * @param ?array<string> $url
+     * @param ?array<string, mixed> $get
+     * @return string
+     */
+    public static function buildFullPath(?array $url, ?array $get): string
     {
         if (empty($url)) {
             $path = '/';
@@ -119,13 +137,16 @@ class Helpers
         return self::removeDuplicateCharacter($path, '\/', '/');
     }
 
-    public static function dateWithTimeFormat($datetime): string
+    public static function dateWithTimeFormat(?string $datetime): string
     {
         if (empty($datetime)) {
             return '';
         }
 
         $timestamp = strtotime($datetime);
+        if ($timestamp === false) {
+            return $datetime;
+        }
 
         $format = [];
         $format[] = 'd.m';
